backend/requirements.txt
# HeyPDF 2.0 — Backend Dependencies
# All versions are pinned to prevent "breaks after a year" issues.
# To install: pip install -r requirements.txt

# ── Web Framework ────────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20        # Required for FastAPI file uploads

# ── PDF Processing ────────────────────────────────────────────────────────────
pdfplumber==0.11.4              # Accurate text + table extraction

# ── Embeddings & Vector Search ────────────────────────────────────────────────
sentence-transformers==3.3.1   # all-MiniLM-L6-v2 local embeddings (~90MB download)
faiss-cpu==1.9.0.post1         # Facebook AI Similarity Search (CPU version)

# ── AI Provider SDKs ─────────────────────────────────────────────────────────
groq==0.13.1                   # Groq API (llama3-8b-8192)
google-generativeai==0.8.3     # Gemini API (gemini-1.5-flash)
openai==1.58.1                 # OpenAI-compatible client (used for OpenRouter)
huggingface-hub==0.27.0        # HuggingFace Inference API

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1           # .env file loading
pydantic==2.10.4               # Data validation (bundled with FastAPI but pinned)
